Create S3MultipartUploadException
