+++ src/Exception/MultipartUploadException.php
@@ -8,8 +8,6 @@ class MultipartUploadException extends \RuntimeException
 {
     /** @var UploadState State of the erroneous transfer */
     private $state;
-    /** @var string File path of transfer object */
-    private $filePath;
 
     /**
      * @param UploadState      $state Upload state at time of the exception.
@@ -24,10 +22,8 @@ class MultipartUploadException extends \RuntimeException
             /** @var $error AwsException */
             foreach ($prev as $part => $error) {
                 $msg .= "- Part {$part}: " . $error->getMessage(). "\n";
-                $this->collectPathInfo($error->getCommand());
             }
         } elseif ($prev instanceof AwsException) {
-            $this->collectPathInfo($prev->getCommand());
             switch ($prev->getCommand()->getName()) {
                 case 'CreateMultipartUpload':
                 case 'InitiateMultipartUpload':
@@ -59,30 +55,5 @@ class MultipartUploadException extends \RuntimeException
     public function getState()
     {
         return $this->state;
-    }
-
-    /**
-     * Get file path of the transfer
-     * Note:
-     * - returns null when the bucket and key could not be
-     *   ascertained from the exception. (e.g., Glacier)
-     *
-     * @return string|null
-     */
-    public function getFilePath()
-    {
-        return $this->filePath;
-    }
-
-    /**
-     * Construct file path
-     *
-     * @param CommandInterface $cmd
-     */
-    private function collectPathInfo(CommandInterface $cmd)
-    {
-        if (empty($this->filePath) && isset($cmd['Bucket']) && isset($cmd['Key'])) {
-            $this->filePath = $cmd['Bucket'] . '/' . $cmd['Key'];
-        }
     }
 }
